* Path-finding, command and unit life-cycle logging re-jigged so it can be filtered (by unit/frame) easily  * command issue checked for duplicate attack commands (AI re-issues attack commands, stresses path-finder on big maps)  * fix for units not waiting a bit on blocked paths (paths were cleared on first block)  * GameConstants::clusterSize increased from 16 to 32...     * reduces avg. search time on 256x256 maps tenfold     * may cause some path quality issues     * NodePool size increased so intra-cluster path searches should all succeed (maybe should test Vs NodeMap)

+++ source/game/game/game.cpp
@@ -441,7 +441,7 @@ void GameState::setScriptDisplay(const string &msg) {
 				++lines;
 			}
 		}
-		m_scriptDisplayPos.y = g_metrics.getScreenH() - 64 - fm->getHeight() * (lines - 1);
+		m_scriptDisplayPos.y = g_metrics.getScreenH() - 64 - int(fm->getHeight() * (lines - 1));
 	}
 }
